Formatting fixes
Fix Application formatting errors

Formatting changes

Ruff autoformat

Add timeout to apple wallet gen

Even more formatting changes

Remove unused linter rules

Remove logging calls

Fix rest of formatting stuff

+++ hiss/shared/management/commands/seeddb.py
@@ -4,12 +4,19 @@ from datetime import timedelta
 from django.core.management import CommandParser, base
 from django.utils import timezone
 
+from application.constants import (
+    CLASSIFICATIONS,
+    GENDERS,
+    GRAD_YEARS,
+    HACKATHON_TIMES,
+    RACES,
+    STATUS_ADMITTED,
+    STATUS_PENDING,
+)
 from application.models import (
     Application,
     Wave,
 )
-from application.constants import GENDERS, RACES, CLASSIFICATIONS, HACKATHON_TIMES, GRAD_YEARS, STATUS_PENDING, \
-    STATUS_ADMITTED
 from user.models import User
 
 COMMON_NAMES = [
@@ -36,7 +43,7 @@ COMMON_NAMES = [
 ]
 
 LAST_NAME = "Seedoe"
-PASSWORD = "password"
+PASSWORD = "password"  # noqa: S105
 
 
 class Command(base.BaseCommand):
@@ -59,14 +66,16 @@ class Command(base.BaseCommand):
         """Function that creates Users and Applications to make testing easier.
         To run use -- python manage.py shell < shared/seed.py
         WARNING - Will delete most of the Users and Applications when run.
-        DO not use in production
+        DO not use in production.
         """
         self.stdout.write("Starting the seeding process.")
         num_users = options["num_users"]
         num_active = options["num_active"]
         num_apps = options["num_applications"]
 
-        assert num_users >= num_active >= num_apps
+        if num_apps > num_active > num_users:
+            msg = "Num users should be greater than num_active which should be greater than num_apps"
+            raise ValueError(msg)
         User.objects.filter(is_superuser=False).delete()
         Application.objects.filter(last_name=LAST_NAME).delete()
         Wave.objects.all().delete()
@@ -81,8 +90,8 @@ class Command(base.BaseCommand):
         for i in range(num_users):
             random_name = random.choice(COMMON_NAMES)
             email = random_name + str(i) + "@seed.com"
-            is_active = True if i < num_active else False
-            will_apply = True if i < num_apps else False
+            is_active = i < num_active
+            will_apply = i < num_apps
             user = User.objects.create_user(email, PASSWORD, is_active=is_active)
             seeds.append((user, will_apply, random_name))
         random.shuffle(seeds)
